refact : Create useOutsideClick custom hook
- useOutsideClick hook dropdown 컴포넌트에서 분리
- Modal 등의 외부 클릭 시, 창이 닫히게 하는데 사용될 것으로 예상되어 분리
